--- src/components/App.js
@@ -1,12 +1,0 @@
-import Home from './Home'
-import { Routes, Route } from 'react-router-dom';
-
-function App() {
-  return (
-    <Routes>
-      <Route path='/' element={<Home />} />
-    </Routes>
-  )
-}
-
-export default App;
