+++ src/components/App/_App.scss
@@ -1,0 +1,66 @@
+@use '../../styles/variables' as *;
+@use '../../styles/mixins' as *;
+
+* {
+  // border: 1px solid red;
+}
+
+html,
+body,
+.root,
+.page {
+  margin: 0;
+  min-height: 100vh;
+}
+
+html {
+  background: linear-gradient(180deg, #D4A29C 0%, #EDCC8B 50%, #B6D8D3 100%);
+  font-size: 62.5%;
+}
+
+h1, h2 {
+  font-family: v.$header-font;
+}
+
+h2 {
+  font-size: 6rem;
+  margin: 1rem 0;
+  align-self: center;
+}
+
+h1,
+h2,
+h3,
+button,
+p {
+  color: v.$text-color;
+  text-shadow: v.$shadow;
+}
+
+h3 {
+  font-family: v.$header-font;
+  font-size: 3rem;
+}
+
+button,
+p {
+  font-size: 3rem;
+}
+
+p {
+  font-family: v.$body-font;
+  margin: 0;
+}
+
+.page {
+  height: 100%;
+  width: 100%;
+}
+
+.page__box {
+  width: 100%;
+  height: 100vh;
+  display: flex;
+  @include m.true-center;
+  background-color: v.$content-box;
+}
