+++ src/components/Footer/_Footer.scss
@@ -1,0 +1,3 @@
+.footer {
+  height: 10rem;
+}
