--- src/components/Home.js
@@ -1,21 +1,0 @@
-import Nav from './Nav'
-import Landing from './Landing'
-import Projects from './Projects'
-import About from './About'
-import Contact from './Contact'
-import Footer from './Footer'
-
-const Home = () => {
-  return (
-    <>
-      <Nav />
-      <Landing />
-      <Projects />
-      <About />
-      <Contact />
-      <Footer />
-    </>
-  )
-}
-
-export default Home;
