+++ src/components/Nav/_Nav.scss
@@ -11,9 +11,20 @@ nav {
   backdrop-filter: blur(5px)
 }
 
+h1 {
+  font-size: 2.6rem;
+}
+
 .nav__text {
   font-size: 2.2rem;
 }
+
+.nav__header,
+.nav__button,
+.nav__button:visited {
+  text-decoration: none;
+}
+
 
 .nav__header,
 .nav__buttons {
@@ -21,6 +32,5 @@ nav {
 }
 
 .nav__button {
-  border: none;
-  padding: 2rem;
+  @include m.button;
 }
