+++ src/styles/output.css
@@ -1,5 +1,4 @@
 .landing {
-  width: 100%;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -15,6 +14,9 @@
 }
 
 .landing__button {
+  font-family: "Amaranth", sans-serif;
+  background: 0%;
+  padding: 0.5em 1em;
   border-radius: 2px;
   border: 1px solid #FFF;
   margin-top: 1em;
@@ -35,8 +37,18 @@ nav {
   backdrop-filter: blur(5px);
 }
 
+h1 {
+  font-size: 2.6rem;
+}
+
 .nav__text {
   font-size: 2.2rem;
+}
+
+.nav__header,
+.nav__button,
+.nav__button:visited {
+  text-decoration: none;
 }
 
 .nav__header,
@@ -45,8 +57,9 @@ nav {
 }
 
 .nav__button {
-  border: none;
-  padding: 2rem;
+  font-family: "Amaranth", sans-serif;
+  background: 0%;
+  padding: 0.5em 1em;
 }
 
 html,
@@ -67,7 +80,7 @@ h1, h2 {
 }
 
 h2 {
-  font-size: 8rem;
+  font-size: 6rem;
   margin: 1rem 0;
   align-self: center;
 }
@@ -81,13 +94,9 @@ p {
   text-shadow: 0px 2px 9px rgba(0, 0, 0, 0.25);
 }
 
-button {
-  font-family: "Amaranth", sans-serif;
-}
-
-p {
-  font-family: "Spinnaker", sans-serif;
-  margin: 0;
+h3 {
+  font-family: "Philosopher", sans-serif;
+  font-size: 3rem;
 }
 
 button,
@@ -95,9 +104,9 @@ p {
   font-size: 3rem;
 }
 
-button {
-  background: 0%;
-  padding: 0.5em 1em;
+p {
+  font-family: "Spinnaker", sans-serif;
+  margin: 0;
 }
 
 .page {
@@ -107,16 +116,11 @@ button {
 
 .page__box {
   width: 100%;
-  height: 60vh;
+  height: 100vh;
   display: flex;
   justify-content: center;
   align-items: center;
   background-color: rgba(162, 99, 96, 0.2509803922);
-}
-
-h3 {
-  font-family: "Philosopher", sans-serif;
-  font-size: 4rem;
 }
 
 /*# sourceMappingURL=output.css.map */
